Schedule-requirements.txt
tweepy
requests
Pillow
beautifulsoup4
pywebpush
deta